requirements.txt
numpy
pandas
pillow
torch
ultralytics
tqdm
seaborn